Update intensity.R
This is a minimal fix for the error reported by CRAN. It seems to be caused when INT only has one row, and is then turned into a vector when it is subset down. Apparently the next version of R will be more pedantic about setting a matrix class on a vector lacking `dim` attributes and errors.